LayoutTests/editing/execCommand/reset-values-after-navigation-expected.txt

PASS queryCommandValues["DefaultParagraphSeparator"] is "div"
PASS queryCommandValues["StyleWithCSS"] is "false"
PASS successfullyParsed is true

TEST COMPLETE
